refactor(bins): change bin id type
